Move Jacobian to dim2 folder
