+++ core/specfem/jacobian.hpp
@@ -1,3 +1,3 @@
 #pragma once
 
-#include "jacobian/jacobian.hpp"
+#include "jacobian/dim2/jacobian.hpp"
